backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
email-validator==2.1.0
paypalrestsdk
paypal-checkout-serversdk
python-dotenv
requests
pyjwt==2.8.0
python-multipart==0.0.6
slowapi==0.1.9
